Remove some ifdefs
