+++ Common/File/VFS/VFS.h
@@ -8,7 +8,16 @@
 // read them manually out of the APK zipfile, while being able to run on other
 // platforms as well with the appropriate directory set-up.
 
-class AssetReader;
+class AssetReader {
+public:
+	virtual ~AssetReader() {}
+	// use delete[]
+	virtual uint8_t *ReadAsset(const char *path, size_t *size) = 0;
+	// Filter support is optional but nice to have
+	virtual bool GetFileListing(const char *path, std::vector<File::FileInfo> *listing, const char *filter = 0) = 0;
+	virtual bool GetFileInfo(const char *path, File::FileInfo *info) = 0;
+	virtual std::string toString() const = 0;
+};
 
 class VFS {
 public:
